rustc_mir_transform

Module check_enums

Structs

CheckEnums ğŸ”’

This pass inserts checks for a valid enum discriminant where they are most likely to find UB, because checking everywhere like Miri would generate too much MIR.

EnumFinder ğŸ”’

A Visitor that finds the construction of enums and evaluates which checks we should apply.

TyAndSize ğŸ”’

Enums

EnumCheckType ğŸ”’

Represent the different kind of enum checks we can insert.

Functions

insert_direct_enum_check ğŸ”’

insert_discr_cast_to_u128 ğŸ”’

Inserts the cast of an operand (any type) to a u128 value that holds the discriminant value.

insert_niche_check ğŸ”’

insert_uninhabited_enum_check ğŸ”’

split_block ğŸ”’