rustc_mir_transform

Module check_pointers

Structs

PointerCheck 🔒

Details of a pointer check, the condition on which we decide whether to fail the assert and an AssertKind that defines the behavior on failure.

PointerFinder 🔒

Enums

BorrowedFieldProjectionMode 🔒

When checking for borrows of field projections (&(*ptr).a), we might want to check for the field type (type of .a in the example). This enum defines the variations (pass the pointer Ty or the field Ty).

Functions

check_pointers 🔒

Utility for adding a check for read/write on every sized, raw pointer.

split_block 🔒