rustc_mir_transform
Module cleanup_post_borrowck

This module provides a pass that removes parts of MIR that are no longer relevant after analysis phase and borrowck. In particular, it removes false edges, user type annotations and replaces following statements with Nops:

AscribeUserType

FakeRead

Assign statements with a Fake borrow

Coverage statements of kind BlockMarker or SpanMarker

Structs

CleanupPostBorrowck 🔒