Struct rustc_mir_transform::const_prop::ConstPropMachine

pub struct ConstPropMachine<'mir, 'tcx> {
    stack: Vec<Frame<'mir, 'tcx>>,
    pub written_only_inside_own_block_locals: FxHashSet<Local>,
    pub only_propagate_inside_block_locals: BitSet<Local>,
    pub can_const_prop: IndexVec<Local, ConstPropMode>,
}

Fields
stack: Vec<Frame<'mir, 'tcx>>

The virtual call stack.

written_only_inside_own_block_locals: FxHashSet<Local>

OnlyInsideOwnBlock locals that were written in the current block get erased at the end.

only_propagate_inside_block_locals: BitSet<Local>

Locals that need to be cleared after every block terminates.

can_const_prop: IndexVec<Local, ConstPropMode>
Implementations

impl ConstPropMachine<'_, '_>

pub fn new(
only_propagate_inside_block_locals: BitSet<Local>,
can_const_prop: IndexVec<Local, ConstPropMode>
) -> Self