rustc_mir_transform::coroutine

Module drop

Drops and async drops related logic for coroutine transformation pass

Structs

FixReturnPendingVisitor 🔒

Functions

build_pin_fut 🔒

build_poll_call 🔒

build_poll_switch 🔒

cleanup_async_drops 🔒

Cleanup all async drops (reset to sync)

create_coroutine_drop_shim 🔒

create_coroutine_drop_shim_async 🔒

create_coroutine_drop_shim_proxy_async 🔒

elaborate_coroutine_drops 🔒

expand_async_drops 🔒

Expand Drop terminator for async drops into mainline poll-switch and dropline poll-switch

gather_dropline_blocks 🔒

has_expandable_async_drops 🔒

insert_clean_drop 🔒