Struct rustc_mir_transform::coverage::counters::CoverageCounters

pub(super) struct CoverageCounters {
    next_counter_id: CounterId,
    next_expression_id: ExpressionId,
    bcb_counters: IndexVec<BasicCoverageBlock, Option<BcbCounter>>,
    bcb_edge_counters: FxHashMap<(BasicCoverageBlock, BasicCoverageBlock), BcbCounter>,
    bcb_has_incoming_edge_counters: BitSet<BasicCoverageBlock>,
    pub(super) intermediate_expressions: Vec<BcbCounter>,
    pub debug_counters: DebugCounters,
}

Generates and stores coverage counter and coverage expression information associated with nodes/edges in the BCB graph.

Fields
next_counter_id: CounterIdnext_expression_id: ExpressionIdbcb_counters: IndexVec<BasicCoverageBlock, Option<BcbCounter>>

Coverage counters/expressions that are associated with individual BCBs.

bcb_edge_counters: FxHashMap<(BasicCoverageBlock, BasicCoverageBlock), BcbCounter>

Coverage counters/expressions that are associated with the control-flow edge between two BCBs.

bcb_has_incoming_edge_counters: BitSet<BasicCoverageBlock>

Tracks which BCBs have a counter associated with some incoming edge. Only used by debug assertions, to verify that BCBs with incoming edge counters do not have their own physical counters (expressions are allowed).

intermediate_expressions: Vec<BcbCounter>

Expression nodes that are not directly associated with any particular BCB/edge, but are needed as operands to more complex expressions. These are always BcbCounter::Expression.

debug_counters: DebugCounters
Implementations

impl CoverageCounters

pub(super) fn new(basic_coverage_blocks: &CoverageGraph) -> Self

pub fn enable_debug(&mut self)

Activate the DebugCounters data structures, to provide additional debug formatting features when formatting BcbCounter (counter) values.

pub fn make_bcb_counters( &mut self, basic_coverage_blocks: &CoverageGraph, coverage_spans: &[CoverageSpan] ) -> Result<(), Error>

Makes BcbCounter Counters and Expressions for the BasicCoverageBlocks directly or indirectly associated with CoverageSpans, and accumulates additional Expressions representing intermediate values.

fn make_counter<F>(&mut self, debug_block_label_fn: F) -> BcbCounterwhere F: Fn() -> Option<String>,

fn make_expression<F>( &mut self, lhs: Operand, op: Op, rhs: Operand, debug_block_label_fn: F ) -> BcbCounterwhere F: Fn() -> Option<String>,

pub fn make_identity_counter(&mut self, counter_operand: Operand) -> BcbCounter

fn next_counter(&mut self) -> CounterId

Counter IDs start from one and go up.

fn next_expression(&mut self) -> ExpressionId

Expression IDs start from 0 and go up. (Counter IDs and Expression IDs are distinguished by the Operand enum.)

fn set_bcb_counter( &mut self, bcb: BasicCoverageBlock, counter_kind: BcbCounter ) -> Result<Operand, Error>

fn set_bcb_edge_counter( &mut self, from_bcb: BasicCoverageBlock, to_bcb: BasicCoverageBlock, counter_kind: BcbCounter ) -> Result<Operand, Error>

pub(super) fn bcb_counter(&self, bcb: BasicCoverageBlock) -> Option<&BcbCounter>

pub(super) fn take_bcb_counter( &mut self, bcb: BasicCoverageBlock ) -> Option<BcbCounter>

pub(super) fn drain_bcb_counters( &mut self ) -> impl Iterator<Item = (BasicCoverageBlock, BcbCounter)> + '_

pub(super) fn drain_bcb_edge_counters( &mut self ) -> impl Iterator<Item = ((BasicCoverageBlock, BasicCoverageBlock), BcbCounter)> + '_