Struct rustc_mir_transform::coverage::debug::UsedExpressions

pub(super) struct UsedExpressions {
    some_used_expression_operands: Option<FxHashMap<Operand, Vec<ExpressionId>>>,
    some_unused_expressions: Option<Vec<(BcbCounter, Option<BasicCoverageBlock>, BasicCoverageBlock)>>,
}

If enabled, this struct captures additional data used to track whether expressions were used, directly or indirectly, to compute the coverage counts for all CoverageSpans, and any that are not used are retained in the unused_expressions Vec, to be included in debug output (logs and/or a CoverageGraph graphviz output).

Fields
some_used_expression_operands: Option<FxHashMap<Operand, Vec<ExpressionId>>>some_unused_expressions: Option<Vec<(BcbCounter, Option<BasicCoverageBlock>, BasicCoverageBlock)>>
Implementations

impl UsedExpressions

pub fn new() -> Self

pub fn enable(&mut self)

pub fn is_enabled(&self) -> bool

pub fn add_expression_operands(&mut self, expression: &BcbCounter)

pub fn expression_is_used(&self, expression: &BcbCounter) -> bool

pub fn add_unused_expression_if_not_found( &mut self, expression: &BcbCounter, edge_from_bcb: Option<BasicCoverageBlock>, target_bcb: BasicCoverageBlock )

pub fn get_unused_expressions( &self ) -> Vec<(BcbCounter, Option<BasicCoverageBlock>, BasicCoverageBlock)>

Return the list of unused counters (if any) as a tuple with the counter (BcbCounter), optional from_bcb (if it was an edge counter), and target_bcb.

pub fn validate( &mut self, bcb_counters_without_direct_coverage_spans: &[(Option<BasicCoverageBlock>, BasicCoverageBlock, BcbCounter)] )

If enabled, validate that every BCB or edge counter not directly associated with a coverage span is at least indirectly associated (it is a dependency of a BCB counter that is associated with a coverage span).

pub fn alert_on_unused_expressions(&self, debug_counters: &DebugCounters)