Struct rustc_mir_transform::coverage::graph::CoverageGraph

pub(crate) struct CoverageGraph {
    bcbs: IndexVec<BasicCoverageBlock, BasicCoverageBlockData>,
    bb_to_bcb: IndexVec<BasicBlock, Option<BasicCoverageBlock>>,
    pub(crate) successors: IndexVec<BasicCoverageBlock, Vec<BasicCoverageBlock>>,
    pub(crate) predecessors: IndexVec<BasicCoverageBlock, Vec<BasicCoverageBlock>>,
    dominators: Option<Dominators<BasicCoverageBlock>>,
}

A coverage-specific simplification of the MIR control flow graph (CFG). The CoverageGraphs nodes are BasicCoverageBlocks, which encompass one or more MIR BasicBlocks.

Fields
bcbs: IndexVec<BasicCoverageBlock, BasicCoverageBlockData>bb_to_bcb: IndexVec<BasicBlock, Option<BasicCoverageBlock>>successors: IndexVec<BasicCoverageBlock, Vec<BasicCoverageBlock>>predecessors: IndexVec<BasicCoverageBlock, Vec<BasicCoverageBlock>>dominators: Option<Dominators<BasicCoverageBlock>>
Implementations

impl CoverageGraph

pub(crate) fn from_mir(mir_body: &Body<'_>) -> Self

fn compute_basic_coverage_blocks( mir_body: &Body<'_>, ) -> (IndexVec<BasicCoverageBlock, BasicCoverageBlockData>, IndexVec<BasicBlock, Option<BasicCoverageBlock>>)

pub(crate) fn iter_enumerated( &self, ) -> impl Iterator<Item = (BasicCoverageBlock, &BasicCoverageBlockData)>

pub(crate) fn bcb_from_bb(&self, bb: BasicBlock) -> Option<BasicCoverageBlock>

pub(crate) fn dominates( &self, dom: BasicCoverageBlock, node: BasicCoverageBlock, ) -> bool

pub(crate) fn cmp_in_dominator_order( &self, a: BasicCoverageBlock, b: BasicCoverageBlock, ) -> Ordering

pub(crate) fn bcb_has_multiple_in_edges(&self, bcb: BasicCoverageBlock) -> bool

Returns true if the given node has 2 or more in-edges, i.e. 2 or more predecessors.

This property is interesting to code that assigns counters to nodes and edges, because if a node doesn’t have multiple in-edges, then there’s no benefit in having a separate counter for its in-edge, because it would have the same value as the node’s own counter.

FIXME: That assumption might not be true for TerminatorKind::Yield?