Module rustc_mir_transform::coverage::spans

Modules

from_mir ğŸ”’

Structs

BcbMapping ğŸ”’

CoverageSpans ğŸ”’

CurrCovspan ğŸ”’

PrevCovspan ğŸ”’

RefinedCovspan ğŸ”’

SpansRefiner ğŸ”’

Converts the initial set of coverage spans (one per MIR Statement or Terminator) into a minimal set of coverage spans, using the BCB CFG to determine where it is safe and useful to:

Enums

BcbMappingKind ğŸ”’

Functions

generate_coverage_spans ğŸ”’

Extracts coverage-relevant spans from MIR, and associates them with their corresponding BCBs.