Function rustc_mir_transform::coverage::spans::function_source_span

pub(super) fn function_source_span(span: Span, body_span: Span) -> Span

Returns an extrapolated span (pre-expansion1) corresponding to a range within the function’s body source. This span is guaranteed to be contained within, or equal to, the body_span. If the extrapolated span is not contained within the body_span, the body_span is returned.

1Expansions result from Rust syntax including macros, syntactic sugar, etc.).

↩