rustc_mir_transform::coverage
Module unexpand

Functions

unexpand_into_body_span 🔒

Walks through the expansion ancestors of original_span to find a span that is contained in body_span and has the same [syntax context] as body_span.

unexpand_into_body_span_with_expn_kind 🔒

Walks through the expansion ancestors of original_span to find a span that is contained in body_span and has the same [syntax context] as body_span.

unexpand_into_body_span_with_prev 🔒

Walks through the expansion ancestors of original_span to find a span that is contained in body_span and has the same syntax context as body_span. The ancestor that was traversed just before the matching span (if any) is also returned.