rustc_mir_transform::deduce_param_attrs

Function deduced_param_attrs

pub(crate) fn deduced_param_attrs<'tcx>(
    tcx: TyCtxt<'tcx>,
    def_id: LocalDefId,
) -> &'tcx [DeducedParamAttrs]

Returns the deduced parameter attributes for a function.

Deduced parameter attributes are those that can only be soundly determined by examining the body of the function instead of just the signature. These can be useful for optimization purposes on a best-effort basis. We compute them here and store them into the crate metadata so dependent crates can use them.