rustc_mir_transform
Module elaborate_drops

Structs

ElaborateDrops 🔒

During MIR building, Drop terminators are inserted in every place where a drop may occur. However, in this phase, the presence of these terminators does not guarantee that a destructor will run, as the target of the drop may be uninitialized. In general, the compiler cannot determine at compile time whether a destructor will run or not.

ElaborateDropsCtxt 🔒

InitializationData 🔒

Functions

compute_dead_unwinds 🔒

Records unwind edges which are known to be unreachable, because they are in drop terminators that can’t drop anything.