Struct rustc_mir_transform::elaborate_drops::ElaborateDropsCtxt

struct ElaborateDropsCtxt<'a, 'tcx> {
    tcx: TyCtxt<'tcx>,
    body: &'a Body<'tcx>,
    env: &'a MoveDataParamEnv<'tcx>,
    init_data: InitializationData<'a, 'tcx>,
    drop_flags: FxHashMap<MovePathIndex, Local>,
    patch: MirPatch<'tcx>,
    un_derefer: UnDerefer<'tcx>,
}

Fields
tcx: TyCtxt<'tcx>body: &'a Body<'tcx>env: &'a MoveDataParamEnv<'tcx>init_data: InitializationData<'a, 'tcx>drop_flags: FxHashMap<MovePathIndex, Local>patch: MirPatch<'tcx>un_derefer: UnDerefer<'tcx>
Implementations

impl<'b, 'tcx> ElaborateDropsCtxt<'b, 'tcx>

fn move_data(&self) -> &'b MoveData<'tcx>

fn param_env(&self) -> ParamEnv<'tcx>

fn create_drop_flag(&mut self, index: MovePathIndex, span: Span)

fn drop_flag(&mut self, index: MovePathIndex) -> Option<Place<'tcx>>

fn elaborate(self) -> MirPatch<'tcx>

create a patch that elaborates all drops in the input MIR.

fn collect_drop_flags(&mut self)

fn elaborate_drops(&mut self)

fn elaborate_replace(
&mut self,
loc: Location,
place: Place<'tcx>,
value: &Operand<'tcx>,
target: BasicBlock,
unwind: Option<BasicBlock>
)

Elaborate a MIR replace terminator. This instruction is not directly handled by codegen, and therefore must be desugared.

The desugaring drops the location if needed, and then writes the value (including setting the drop flag) over it in both arms.

The replace terminator can also be called on places that are not tracked by elaboration (for example, replace x[i] <- tmp0). The borrow checker requires that these locations are initialized before the assignment, so we just generate an unconditional drop.

fn constant_bool(&self, span: Span, val: bool) -> Rvalue<'tcx>

fn set_drop_flag(&mut self, loc: Location, path: MovePathIndex, val: DropFlagState)

fn drop_flags_on_init(&mut self)

fn drop_flags_for_fn_rets(&mut self)

fn drop_flags_for_args(&mut self)

fn drop_flags_for_locs(&mut self)