rustc_mir_transform

Function mir_drops_elaborated_and_const_checked

pub(crate) fn mir_drops_elaborated_and_const_checked(
    tcx: TyCtxt<'_>,
    def: LocalDefId,
) -> &Steal<Body<'_>>

Obtain just the main MIR (no promoteds) and run some cleanups on it. This also runs mir borrowck before doing so in order to ensure that borrowck can be run and doesn’t end up missing the source MIR due to stealing happening.