Function rustc_mir_transform::promoted_mir

pub(crate) fn promoted_mir<'tcx>(
    tcx: TyCtxt<'tcx>,
    def: WithOptConstParam<LocalDefId>
) -> &'tcx IndexVec<Promoted, Body<'tcx>>

Fetch all the promoteds of an item and prepare their MIR bodies to be ready for constant evaluation once all substitutions become known.