Function rustc_mir_transform::generator::compute_storage_conflicts

fn compute_storage_conflicts<'mir, 'tcx>(
    body: &'mir Body<'tcx>,
    saved_locals: &GeneratorSavedLocals,
    always_live_locals: BitSet<Local>,
    requires_storage: Results<'tcx, MaybeRequiresStorage<'mir, 'tcx>>
) -> BitMatrix<GeneratorSavedLocal, GeneratorSavedLocal>

For every saved local, looks for which locals are StorageLive at the same time. Generates a bitset for every local of all the other locals that may be StorageLive simultaneously with that local. This is used in the layout computation; see GeneratorLayout for more.