Struct rustc_mir_transform::generator::GeneratorSavedLocals

struct GeneratorSavedLocals(BitSet<Local>);

The set of Locals that must be saved across yield points.

GeneratorSavedLocal is indexed in terms of the elements in this set; i.e. GeneratorSavedLocal::new(1) corresponds to the second local included in this set.

Tuple Fields
0: BitSet<Local>
Implementations

impl GeneratorSavedLocals

fn iter_enumerated(
&self
) -> impl '_ + Iterator<Item = (GeneratorSavedLocal, Local)>

Returns an iterator over each GeneratorSavedLocal along with the Local it corresponds to.

fn renumber_bitset(&self, input: &BitSet<Local>) -> BitSet<GeneratorSavedLocal>

Transforms a BitSet<Local> that contains only locals saved across yield points to the equivalent BitSet<GeneratorSavedLocal>.

fn get(&self, local: Local) -> Option<GeneratorSavedLocal>