Enum rustc_mir_transform::gvn::Value

enum Value<'tcx> {

    Opaque(usize),
    Constant(Const<'tcx>),
    Aggregate(Ty<'tcx>, VariantIdx, Vec<VnIndex>),
    Repeat(VnIndex, Const<'tcx>),
    Address {
        place: Place<'tcx>,
        provenance: usize,
    },
    Projection(VnIndex, ProjectionElem<VnIndex, Ty<'tcx>>),
    Discriminant(VnIndex),
    Len(VnIndex),
    NullaryOp(NullOp<'tcx>, Ty<'tcx>),
    UnaryOp(UnOp, VnIndex),
    BinaryOp(BinOp, VnIndex, VnIndex),
    CheckedBinaryOp(BinOp, VnIndex, VnIndex),
    Cast {
        kind: CastKind,
        value: VnIndex,
        from: Ty<'tcx>,
        to: Ty<'tcx>,
    },

}

Variants

Opaque(usize)

Used to represent values we know nothing about. The usize is a counter incremented by new_opaque.

Constant(Const<'tcx>)

Evaluated or unevaluated constant value.

Aggregate(Ty<'tcx>, VariantIdx, Vec<VnIndex>)

An aggregate value, either tuple/closure/struct/enum. This does not contain unions, as we cannot reason with the value.

Repeat(VnIndex, Const<'tcx>)

This corresponds to a [value; count] expression.

Address

Fields

place: Place<'tcx>

provenance: usize

Give each borrow and pointer a different provenance, so we don’t merge them.

The address of a place.

Projection(VnIndex, ProjectionElem<VnIndex, Ty<'tcx>>)

This is the value obtained by projecting another value.

Discriminant(VnIndex)

Discriminant of the given value.

Len(VnIndex)

Length of an array or slice.

NullaryOp(NullOp<'tcx>, Ty<'tcx>)

UnaryOp(UnOp, VnIndex)

BinaryOp(BinOp, VnIndex, VnIndex)

CheckedBinaryOp(BinOp, VnIndex, VnIndex)

Cast

Fields

kind: CastKind

value: VnIndex

from: Ty<'tcx>

to: Ty<'tcx>