rustc_mir_transform::known_panics_lint
Struct ConstPropagator

struct ConstPropagator<'mir, 'tcx> {
    ecx: InterpCx<'tcx, DummyMachine>,
    tcx: TyCtxt<'tcx>,
    param_env: ParamEnv<'tcx>,
    worklist: Vec<BasicBlock>,
    visited_blocks: BitSet<BasicBlock>,
    locals: IndexVec<Local, Value<'tcx>>,
    body: &'mir Body<'tcx>,
    written_only_inside_own_block_locals: FxHashSet<Local>,
    can_const_prop: IndexVec<Local, ConstPropMode>,
}

Visits MIR nodes, performs const propagation and runs lint checks as it goes

Fields
ecx: InterpCx<'tcx, DummyMachine>tcx: TyCtxt<'tcx>param_env: ParamEnv<'tcx>worklist: Vec<BasicBlock>visited_blocks: BitSet<BasicBlock>locals: IndexVec<Local, Value<'tcx>>body: &'mir Body<'tcx>written_only_inside_own_block_locals: FxHashSet<Local>can_const_prop: IndexVec<Local, ConstPropMode>
Implementations

impl<'mir, 'tcx> ConstPropagator<'mir, 'tcx>

fn new(body: &'mir Body<'tcx>, tcx: TyCtxt<'tcx>) -> ConstPropagator<'mir, 'tcx>

fn local_decls(&self) -> &'mir LocalDecls<'tcx>

fn get_const(&self, place: Place<'tcx>) -> Option<&Value<'tcx>>

fn remove_const(&mut self, local: Local)

Remove local from the pool of Locals. Allows writing to them, but not reading from them anymore.

fn access_mut(&mut self, place: &Place<'_>) -> Option<&mut Value<'tcx>>

fn lint_root(&self, source_info: SourceInfo) -> Option<HirId>

fn use_ecx<F, T>(&mut self, f: F) -> Option<T>
where F: FnOnce(&mut Self) -> InterpResult<'tcx, T>,

fn eval_constant(&mut self, c: &ConstOperand<'tcx>) -> Option<ImmTy<'tcx>>

Returns the value, if any, of evaluating c.

fn eval_place(&mut self, place: Place<'tcx>) -> Option<ImmTy<'tcx>>

Returns the value, if any, of evaluating place.

fn eval_operand(&mut self, op: &Operand<'tcx>) -> Option<ImmTy<'tcx>>

Returns the value, if any, of evaluating op. Calls upon eval_constant or eval_place, depending on the variant of Operand used.

fn report_assert_as_lint( &self, location: Location, lint_kind: AssertLintKind, assert_kind: AssertKind<impl Debug>, )

fn check_unary_op( &mut self, op: UnOp, arg: &Operand<'tcx>, location: Location, ) -> Option<()>

fn check_binary_op( &mut self, op: BinOp, left: &Operand<'tcx>, right: &Operand<'tcx>, location: Location, ) -> Option<()>

fn check_rvalue( &mut self, rvalue: &Rvalue<'tcx>, location: Location, ) -> Option<()>

fn check_assertion( &mut self, expected: bool, msg: &AssertKind<Operand<'tcx>>, cond: &Operand<'tcx>, location: Location, )

fn ensure_not_propagated(&self, local: Local)

fn eval_rvalue( &mut self, rvalue: &Rvalue<'tcx>, dest: &Place<'tcx>, ) -> Option<()>