rustc_mir_transform

Module lint_tail_expr_drop_order

Structs

DestructorLabel 🔒

DropsReachable 🔒

LocalLabel 🔒

TailExprDropOrderLint 🔒

Enums

MovePathIndexAtBlock 🔒

Cache entry of drop at a BasicBlock

Functions

assign_observables_names 🔒

Assign names for anonymous or temporary values for diagnosis

collect_user_names 🔒

Extract binding names if available for diagnosis

place_descendent_of_bids 🔒

Check if a moved place at idx is a part of a BID. The use of this check is that we will consider drops on these as a drop of the overall BID and, thus, we can exclude it from the diagnosis.

place_has_common_prefix 🔒

run_lint 🔒

The core of the lint tail-expr-drop-order