rustc_mir_transform::lint_tail_expr_drop_order
Function place_descendent_of_bids

fn place_descendent_of_bids<'tcx>(
    idx: MovePathIndex,
    move_data: &MoveData<'tcx>,
    bids: &UnordSet<&Place<'tcx>>,
) -> bool

Check if a moved place at idx is a part of a BID. The use of this check is that we will consider drops on these as a drop of the overall BID and, thus, we can exclude it from the diagnosis.