rustc_mir_transform::lint_tail_expr_drop_order
Function ty_dtor_span

fn ty_dtor_span<'tcx>(tcx: TyCtxt<'tcx>, ty: Ty<'tcx>) -> Option<Span>

Extract the span of the custom destructor of a type especially the span of the impl Drop header or its entire block when we are working with current local crate.