rustc_mir_transform::match_branches
Trait SimplifyMatch

trait SimplifyMatch<'tcx> {
    // Required methods
    fn can_simplify(
        &mut self,
        tcx: TyCtxt<'tcx>,
        targets: &SwitchTargets,
        typing_env: TypingEnv<'tcx>,
        bbs: &IndexSlice<BasicBlock, BasicBlockData<'tcx>>,
        discr_ty: Ty<'tcx>,
    ) -> Option<()>;
    fn new_stmts(
        &self,
        tcx: TyCtxt<'tcx>,
        targets: &SwitchTargets,
        typing_env: TypingEnv<'tcx>,
        patch: &mut MirPatch<'tcx>,
        parent_end: Location,
        bbs: &IndexSlice<BasicBlock, BasicBlockData<'tcx>>,
        discr_local: Local,
        discr_ty: Ty<'tcx>,
    );

    // Provided method
    fn simplify(
        &mut self,
        tcx: TyCtxt<'tcx>,
        body: &mut Body<'tcx>,
        switch_bb_idx: BasicBlock,
        typing_env: TypingEnv<'tcx>,
    ) -> Option<()> { ... }
}

Required Methods

fn can_simplify( &mut self, tcx: TyCtxt<'tcx>, targets: &SwitchTargets, typing_env: TypingEnv<'tcx>, bbs: &IndexSlice<BasicBlock, BasicBlockData<'tcx>>, discr_ty: Ty<'tcx>, ) -> Option<()>

Check that the BBs to be simplified satisfies all distinct and that the terminator are the same. There are also conditions for different ways of simplification.

fn new_stmts( &self, tcx: TyCtxt<'tcx>, targets: &SwitchTargets, typing_env: TypingEnv<'tcx>, patch: &mut MirPatch<'tcx>, parent_end: Location, bbs: &IndexSlice<BasicBlock, BasicBlockData<'tcx>>, discr_local: Local, discr_ty: Ty<'tcx>, )

Provided Methods

fn simplify( &mut self, tcx: TyCtxt<'tcx>, body: &mut Body<'tcx>, switch_bb_idx: BasicBlock, typing_env: TypingEnv<'tcx>, ) -> Option<()>

Simplifies a match statement, returning true if the simplification succeeds, false otherwise. Generic code is written here, and we generally don’t need a custom implementation.