rustc_mir_transform

Module prettify

These two passes provide no value to the compiler, so are off at every level.

However, they can be enabled on the command line (-Zmir-enable-passes=+ReorderBasicBlocks,+ReorderLocals) to make the MIR easier to read for humans.

Structs

BasicBlockUpdater 🔒

LocalFinder 🔒

LocalUpdater 🔒

ReorderBasicBlocks 🔒

Rearranges the basic blocks into a reverse post-order.

ReorderLocals 🔒

Rearranges the locals into use order.

Functions

permute 🔒