rustc_mir_transform

Module ref_prop

Structs

ReferencePropagation 🔒

Propagate references using SSA analysis.

Replacer 🔒

Utility to help performing substitution of *pattern by target.

Enums

Value 🔒

Functions

compute_replacement 🔒

For each local, save the place corresponding to *local.

fully_replaceable_locals 🔒

Compute the set of locals that can be fully replaced.

move_to_copy_pointers 🔒

The SSA analysis done by SsaLocals treats Operand::Move as a read, even though in general Operand::Move represents pass-by-pointer where the callee can overwrite the pointee (Miri always considers the place deinitialized). CopyProp has a similar trick to turn Operand::Move into Operand::Copy when required for an optimization, but in this pass we just turn all moves of pointers into copies because pointers should be by-value anyway.

propagate_ssa 🔒