Module rustc_mir_transform::unreachable_enum_branching

A pass that eliminates branches on uninhabited or unreachable enum variants.

Structs

UnreachableEnumBranching

Functions

get_discriminant_local 🔒

get_switched_on_type 🔒

If the basic block terminates by switching on a discriminant, this returns the Ty the discriminant is read from. Otherwise, returns None.

variant_discriminants 🔒