rustc_mir_transform
Module add_moves_for_packed_drops

Structs

AddMovesForPackedDrops 🔒

This pass moves values being dropped that are within a packed struct to a separate local before dropping them, to ensure that they are dropped from an aligned address.

Functions

add_move_for_packed_drop 🔒