Module rustc_mir_transform::coverage

Modules

counters ğŸ”’

debug ğŸ”’

The InstrumentCoverage MIR pass implementation includes debugging tools and options to help developers understand and/or improve the analysis and instrumentation of a MIR.

graph ğŸ”’

query

spans ğŸ”’

Structs

Error ğŸ”’

A simple error message wrapper for coverage::Errors.

InstrumentCoverage

Inserts StatementKind::Coverage statements that either instrument the binary with injected counters, via intrinsic llvm.instrprof.increment, and/or inject metadata used during codegen to construct the coverage map.

Instrumentor ğŸ”’

Functions

fn_sig_and_body ğŸ”’

get_body_span ğŸ”’

hash_mir_source ğŸ”’

inject_edge_counter_basic_block ğŸ”’

inject_intermediate_expression ğŸ”’

inject_statement ğŸ”’

make_code_region ğŸ”’

Convert the Span into its file name, start line and column, and end line and column