rustc_mir_transform
Module coverage

Modules

counters ğŸ”’

graph ğŸ”’

mappings ğŸ”’

query ğŸ”’

spans ğŸ”’

unexpand ğŸ”’

Structs

ExtractedHirInfo ğŸ”’

Function information extracted from HIR by the coverage instrumentor.

InstrumentCoverage ğŸ”’

Inserts StatementKind::Coverage statements that either instrument the binary with injected counters, via intrinsic llvm.instrprof.increment, and/or inject metadata used during codegen to construct the coverage map.

Functions

check_source_region ğŸ”’

If llvm-cov sees a source region that is improperly ordered (end < start), it will immediately exit with a fatal error. To prevent that from happening, discard regions that are improperly ordered, or might be interpreted in a way that makes them improperly ordered.

create_mappings ğŸ”’

For each coverage span extracted from MIR, create a corresponding mapping.

extract_hir_info ğŸ”’

extract_hole_spans_from_hir ğŸ”’

hash_mir_source ğŸ”’

inject_coverage_statements ğŸ”’

For each BCB node or BCB edge that has an associated coverage counter, inject any necessary coverage statements into MIR.

inject_edge_counter_basic_block ğŸ”’

Given two basic blocks that have a control-flow edge between them, creates and returns a new block that sits between those blocks.

inject_mcdc_statements ğŸ”’

For each conditions inject statements to update condition bitmap after it has been evaluated. For each decision inject statements to update test vector bitmap after it has been evaluated.

inject_statement ğŸ”’

instrument_function_for_coverage ğŸ”’

make_source_region ğŸ”’

Convert the Span into its file name, start line and column, and end line and column.