rustc_mir_transform

Module coverage

Modules

counters ğŸ”’

graph ğŸ”’

hir_info ğŸ”’

mappings ğŸ”’

query ğŸ”’

spans ğŸ”’

unexpand ğŸ”’

Structs

InstrumentCoverage ğŸ”’

Inserts StatementKind::Coverage statements that either instrument the binary with injected counters, via intrinsic llvm.instrprof.increment, and/or inject metadata used during codegen to construct the coverage map.

Functions

create_mappings ğŸ”’

For each coverage span extracted from MIR, create a corresponding mapping.

inject_coverage_statements ğŸ”’

Inject any necessary coverage statements into MIR, so that they influence codegen.

inject_statement ğŸ”’

instrument_function_for_coverage ğŸ”’