rustc_mir_transform::coverage::counters
Struct CoverageCounters

pub(super) struct CoverageCounters {
    counter_increment_sites: IndexVec<CounterId, CounterIncrementSite>,
    bcb_counters: IndexVec<BasicCoverageBlock, Option<BcbCounter>>,
    bcb_edge_counters: FxHashMap<(BasicCoverageBlock, BasicCoverageBlock), BcbCounter>,
    expressions: IndexVec<ExpressionId, BcbExpression>,
    expressions_memo: FxHashMap<BcbExpression, BcbCounter>,
}

Generates and stores coverage counter and coverage expression information associated with nodes/edges in the BCB graph.

Fields
counter_increment_sites: IndexVec<CounterId, CounterIncrementSite>

List of places where a counter-increment statement should be injected into MIR, each with its corresponding counter ID.

bcb_counters: IndexVec<BasicCoverageBlock, Option<BcbCounter>>

Coverage counters/expressions that are associated with individual BCBs.

bcb_edge_counters: FxHashMap<(BasicCoverageBlock, BasicCoverageBlock), BcbCounter>

Coverage counters/expressions that are associated with the control-flow edge between two BCBs.

We currently don’t iterate over this map, but if we do in the future, switch it back to FxIndexMap to avoid query stability hazards.

expressions: IndexVec<ExpressionId, BcbExpression>

Table of expression data, associating each expression ID with its corresponding operator (+ or -) and its LHS/RHS operands.

expressions_memo: FxHashMap<BcbExpression, BcbCounter>

Remember expressions that have already been created (or simplified), so that we don’t create unnecessary duplicates.

Implementations

impl CoverageCounters

pub(super) fn make_bcb_counters( basic_coverage_blocks: &CoverageGraph, bcb_needs_counter: &BitSet<BasicCoverageBlock>, ) -> Self

Ensures that each BCB node needing a counter has one, by creating physical counters or counter expressions for nodes and edges as required.

fn with_num_bcbs(num_bcbs: usize) -> Self

fn make_counter_inner(&mut self, site: CounterIncrementSite) -> BcbCounter

Shared helper used by Self::make_phys_node_counter and Self::make_phys_edge_counter. Don’t call this directly.

fn make_phys_node_counter(&mut self, bcb: BasicCoverageBlock) -> BcbCounter

Creates a new physical counter attached a BCB node. The node must not already have a counter.

fn make_phys_edge_counter( &mut self, from_bcb: BasicCoverageBlock, to_bcb: BasicCoverageBlock, ) -> BcbCounter

Creates a new physical counter attached to a BCB edge. The edge must not already have a counter.

fn make_expression( &mut self, lhs: BcbCounter, op: Op, rhs: BcbCounter, ) -> BcbCounter

fn make_expression_inner( expressions: &mut IndexVec<ExpressionId, BcbExpression>, new_expr: BcbExpression, ) -> BcbCounter

This is an associated function so that we can call it while borrowing &mut self.expressions_memo.

fn make_sum(&mut self, counters: &[BcbCounter]) -> Option<BcbCounter>

Creates a counter that is the sum of the given counters.

Returns None if the given list of counters was empty.

pub(super) fn num_counters(&self) -> usize

fn set_bcb_counter( &mut self, bcb: BasicCoverageBlock, counter_kind: BcbCounter, ) -> BcbCounter

fn set_bcb_edge_counter( &mut self, from_bcb: BasicCoverageBlock, to_bcb: BasicCoverageBlock, counter_kind: BcbCounter, ) -> BcbCounter

pub(super) fn term_for_bcb(&self, bcb: BasicCoverageBlock) -> Option<CovTerm>

pub(super) fn counter_increment_sites( &self, ) -> impl Iterator<Item = (CounterId, &CounterIncrementSite)>

Returns an iterator over all the nodes/edges in the coverage graph that should have a counter-increment statement injected into MIR, along with each site’s corresponding counter ID.

pub(super) fn bcb_nodes_with_coverage_expressions( &self, ) -> impl Iterator<Item = (BasicCoverageBlock, ExpressionId)> + Captures<'_>

Returns an iterator over the subset of BCB nodes that have been associated with a counter expression, along with the ID of that expression.

pub(super) fn into_expressions(self) -> IndexVec<ExpressionId, Expression>