Module rustc_mir_transform::coverage::graph

Structs

BasicCoverageBlock 🔒

A node in the control-flow graph of CoverageGraph.

BasicCoverageBlockData 🔒

BasicCoverageBlockData holds the data indexed by a BasicCoverageBlock.

CoverageGraph 🔒

A coverage-specific simplification of the MIR control flow graph (CFG). The CoverageGraphs nodes are BasicCoverageBlocks, which encompass one or more MIR BasicBlocks.

TraversalContext 🔒

Maintains separate worklists for each loop in the BasicCoverageBlock CFG, plus one for the CoverageGraph outside all loops. This supports traversing the BCB CFG in a way that ensures a loop is completely traversed before processing Blocks after the end of the loop.

TraverseCoverageGraphWithLoops 🔒

Enums

CoverageSuccessors 🔒

Holds the coverage-relevant successors of a basic block’s terminator, and indicates whether that block can potentially be combined into the same BCB as its sole successor.

Constants

START_BCB 🔒

Functions

bcb_filtered_successors 🔒

find_loop_backedges 🔒

short_circuit_preorder 🔒

Type Aliases

BcbSuccessors 🔒

Opaque Types

🔒