rustc_mir_transform::coverage
Module mappings

Structs

BranchPair 🔒

This is separate from MCDCBranch to help prepare for larger changes that will be needed for improved branch coverage in the future. (See https://github.com/rust-lang/rust/pull/124217.)

CodeMapping 🔒

Associates an ordinary executable code span with its corresponding BCB.

ExtractedMappings 🔒

MCDCBranch 🔒

Associates an MC/DC branch span with condition info besides fields for normal branch.

MCDCDecision 🔒

Associates an MC/DC decision with its join BCBs.

Constants

MCDC_MAX_BITMAP_SIZE 🔒

Functions

calc_test_vectors_index 🔒

extract_all_mapping_info_from_mir 🔒

Extracts coverage-relevant spans from MIR, and associates them with their corresponding BCBs.

extract_branch_pairs 🔒

extract_mcdc_mappings 🔒

resolve_block_markers 🔒