Module rustc_mir_transform::coverage::query

Structs

CoverageVisitor 🔒

The num_counters argument to llvm.instrprof.increment is the max counter_id + 1, or in other words, the number of counter value references injected into the MIR (plus 1 for the reserved ZERO counter, which uses counter ID 0 when included in an expression). Injected counters have a counter ID from 1..num_counters-1.

Functions

coverageinfo 🔒

covered_code_regions 🔒

is_inlined 🔒

mir_body 🔒

This function ensures we obtain the correct MIR for the given item irrespective of whether that means const mir or runtime mir. For const fn this opts for runtime mir.

provide 🔒

A query provider for retrieving coverage information injected into MIR.