rustc_mir_transform::coverage
Module query

Functions

all_coverage_in_mir_body 🔒

coverage_attr_on 🔒

Query implementation for coverage_attr_on.

coverage_ids_info 🔒

Query implementation for coverage_ids_info.

identify_zero_expressions 🔒

Identify expressions that will always have a value of zero, and note their IDs in a BitSet. Mappings that refer to a zero expression can instead become mappings to a constant zero value.

is_eligible_for_coverage 🔒

Hook implementation for TyCtxt::is_eligible_for_coverage.

is_inlined 🔒

is_zero_term 🔒

Returns true if the given term is known to have a value of zero, taking into account knowledge of which counters are unused and which expressions are always zero.

provide 🔒

Registers query/hook implementations related to coverage.