Module rustc_mir_transform::coverage::query

Structs

CoverageVisitor 🔒

Coverage codegen needs to know the total number of counter IDs and expression IDs that have been used by a function’s coverage mappings. These totals are used to create vectors to hold the relevant counter and expression data, and the maximum counter ID (+ 1) is also needed by the llvm.instrprof.increment intrinsic.

Functions

all_coverage_in_mir_body 🔒

coverageinfo 🔒

covered_code_regions 🔒

is_inlined 🔒

mir_body 🔒

This function ensures we obtain the correct MIR for the given item irrespective of whether that means const mir or runtime mir. For const fn this opts for runtime mir.

provide 🔒

A query provider for retrieving coverage information injected into MIR.