rustc_mir_transform::coverage::spans
Module from_mir

Structs

ExtractedCovspans ğŸ”’

Hole ğŸ”’

SpanFromMir ğŸ”’

Functions

bcb_to_initial_coverage_spans ğŸ”’

extract_covspans_from_mir ğŸ”’

Traverses the MIR body to produce an initial collection of coverage-relevant spans, each associated with a node in the coverage graph (BCB) and possibly other metadata.

filtered_statement_span ğŸ”’

If the MIR Statement has a span contributive to computing coverage spans, return it; otherwise return None.

filtered_terminator_span ğŸ”’

If the MIR Terminator has a span contributive to computing coverage spans, return it; otherwise return None.