rustc_mir_transform
Module deduplicate_blocks

This pass finds basic blocks that are completely equal, and replaces all uses with just one of them.

Structs

BasicBlockHashable ğŸ”’

DeduplicateBlocks ğŸ”’

OptApplier ğŸ”’

Functions

find_duplicates ğŸ”’

hash_statements ğŸ”’

operand_eq ğŸ”’

operand_hash ğŸ”’

rvalue_eq ğŸ”’

rvalue_hash ğŸ”’

statement_eq ğŸ”’

statement_hash ğŸ”’