rustc_mir_transform
Module pass_manager

Structs

Lint ğŸ”’

An adapter for MirLints that implements MirPass.

WithMinOptLevel ğŸ”’

Constants

PASS_TO_PROFILER_NAMES ğŸ”’

Maps MIR pass names to a snake case form to match profiling naming style

Traits

MirLint ğŸ”’

Just like MirPass, except it cannot mutate Body, and MIR dumping is disabled (via the Lint adapter).

MirPass ğŸ”’

A streamlined trait that you can implement to create a pass; the pass will be named after the type, and it will consist of a main loop that goes over each available MIR and applies run_pass.

Functions

c_name ğŸ”’

dump_mir_for_pass ğŸ”’

dump_mir_for_phase_change ğŸ”’

run_passes ğŸ”’

The optional phase_change is applied after executing all the passes, if present

run_passes_inner ğŸ”’

run_passes_no_validate ğŸ”’

Run the sequence of passes without validating the MIR after each pass. The MIR is still validated at the end.

should_run_pass ğŸ”’

to_profiler_name ğŸ”’

Converts a MIR pass name into a snake case form to match the profiling naming style.

validate_body ğŸ”’