Module rustc_mir_transform::pass_manager

Structs

Lint

An adapter for MirLints that implements MirPass.

WithMinOptLevel

Traits

MirLint

Just like MirPass, except it cannot mutate Body.

Functions

dump_mir_for_pass

dump_mir_for_phase_change

run_passes

The optional phase_change is applied after executing all the passes, if present

run_passes_inner 🔒

run_passes_no_validate

Run the sequence of passes without validating the MIR after each pass. The MIR is still validated at the end.

should_run_pass

validate_body