rustc_mir_transform

Module ref_prop

Structs

ReferencePropagation 🔒

Propagate references using SSA analysis.

Replacer 🔒

Utility to help performing substitution of *pattern by target.

Enums

Value 🔒

Functions

compute_replacement 🔒

For each local, save the place corresponding to *local.

fully_replaceable_locals 🔒

Compute the set of locals that can be fully replaced.

propagate_ssa 🔒